rename: 컴포넌트의 맴리스트 파일명 수정, 이동
