+++ src/components/mapPropertyListDialog/ToggleCompare.tsx
